requirements.txt
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyYAML==6.0.1
requests==2.31.0
# SQLite is included in Python standard library


